+++ db/migrate/20190611094802_create_side_effects.rb
@@ -1,7 +1,7 @@
 class CreateSideEffects < ActiveRecord::Migration[5.2]
   def change
     create_table    :side_effects, :primary_key => :side_effect_id do |t|
-    	t.integer        :encounter_id
+    	t.bigint        :encounter_id
     	t.integer      :concept_id
     	t.integer       :value_coded
     	t.boolean     :voided, null: false, default: 0
