+++ db/migrate/20190611095636_create_potential_duplicates.rb
@@ -1,8 +1,8 @@
 class CreatePotentialDuplicates < ActiveRecord::Migration[5.2]
   def change
     create_table    :potential_duplicates, :primary_key => :potential_duplicate_id do |t|
-    	t.integer            :person_id_a, null: false
-    	t.integer            :person_id_b, null: false
+    	t.bigint            :person_id_a, null: false
+    	t.bigint            :person_id_b, null: false
     	t.integer          :duplicate_status_id, null: false
     	t.float             :score, null: false
         t.bigint            :creator, null: false
