+++ db/migrate/20190612082436_create_guardians.rb
@@ -1,10 +1,10 @@
 class CreateGuardians < ActiveRecord::Migration[5.2]
   def change
     create_table :guardians, :primary_key => :guardian_id do |t|
-      t.integer        :person_id, null:false
-      t.integer      :person_a, null:false
-      t.integer      :person_b, null:false
-      t.integer      :relationship_type_id, null:false
+      t.bigint        :person_id, null:false
+      t.bigint      :person_a, null:false
+      t.bigint      :person_b, null:false
+      t.bigint      :relationship_type_id, null:false
       t.boolean     :voided, null:false, default: 0
       t.bigint        :creator
       t.bigint         :voided_by
